// frontend/src/lib/generated/Incentives.types.ts
export type Uint128 = string;
export type Addr = string;
export type AssetInfo = {
  token: {
    contract_addr: Addr;
  };
} | {
  native_token: {
    denom: string;
  };
};
// ...
export interface Asset {
  amount: Uint128;
  info: AssetInfo;
}
// ...
export type Decimal256 = string;
export type ArrayOfScheduleResponse = ScheduleResponse[];
export interface ScheduleResponse {
  end_ts: number;
  rps: Decimal256;
  start_ts: number;
}
export type Boolean = boolean;
export type ArrayOfString = string[];
export type ArrayOfAsset = Asset[];
export type RewardType = {
  int: AssetInfo;
} | {
  ext: {
    info: AssetInfo;
    next_update_ts: number;
  };
};
export interface PoolInfoResponse {
  last_update_ts: number;
  rewards: RewardInfo[];
  total_lp: Uint128;
}
export interface RewardInfo {
  index: Decimal256;
  orphaned: Decimal256;
  reward: RewardType;
  rps: Decimal256;
}